[dev] Unicode characters recognised in classes
Classes are now always checked on the two next code units, and
they are prepared with the library regenerate to compute the
ranges in terms of surrogates. When it can be statically
determined if there is one or two surrogate, currPos is statically
increased of this number; else it depends on execution.

A new option is added during generation: "unicode", which can be:
* false: no surrogate handling
* true: well-formed Unicode accepted
* "lone-surrogates": well-formed Unicode or lone surrogates
    accepted as fallback

Although this option could be added in part for the parser user
(e.g. lone surrogates accepted or not), it would be more complex
and I guess the grammar author should remain the master about
such subtleties concerning her/his grammar.

Various conditions are enforced on classes to avoid edge cases
for this patch:
* a class can contain only of these types: 1. high surrogates,
    2. low surrogates, 3. well-formed characters,
* astral characters cannot be used without the unicode option,
* surrogates cannot be used when option unicode is true
* surrogates cannot be used with the ignoreCase flag or in an
    inverted class

Three opcodes are added:
* MATCH_ANY_CHARACTER: similar to MATCH_ANY, but match a well-formed
    Unicode character, possibly with 2 code units
* ACCEPT_CHARACTER: accept one Unicode character, should be used
    after MATCH_ANY_CHARACTER
* ACCEPT_LONE_CHARACTER: accept one Unicode character or a lone
    surrogate as fallback, can be used after MATCH_ANY

I checked the generated regexes for the Unicode classes used in PEG.js
grammar, and the only difference is for the trivial ranges \u{N}-\u{N+1}
transformed to \u{N}\u{N+1}.

Issue: #586