[dev] Unicode characters recognised in classes (bis)
Compared to 853063d36948e21b5a438cfd5d7a8fea45c0484a this version
has the 'dot' rule which always captures 1 code unit, which is
more quick. To compensate this, the only place where a treatment
is needed in the characters classes to reconstruct the Unicode
characters.

In this version there is a global flag options.unicode, which
outputs a parser where the 'dot' rule and the characters classes
are Unicode-enabled.

I am not sure if Unicode characters classes should be authorised
when options.unicode is false, or possibly if a specific flag should
be added like unicodeCharactersClass to recognise Unicode characters
classes in a non-unicode parser (like the grammar of PEG.js grammars).